requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0.76
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyyaml>=6.0.0
mediapipe>=0.10.0
